# [Interp] Create an intrinsic for Interlocked.Exchange for int64