--- a/mono/metadata/threads-types.h
+++ b/mono/metadata/threads-types.h
@@ -580,4 +580,17 @@ mono_threads_summarize_execute (MonoContext *ctx, gchar **out, MonoStackHash *ha
 gboolean
 mono_threads_summarize_one (MonoThreadSummary *out, MonoContext *ctx);
 
+#if SIZEOF_VOID_P == 4
+/* Spin lock for unaligned InterlockedXXX 64 bit functions on 32bit platforms. */
+extern mono_mutex_t mono_interlocked_mutex;
+static inline void
+mono_interlocked_lock(void) { 
+	mono_os_mutex_lock (&mono_interlocked_mutex);
+}
+static inline void
+mono_interlocked_unlock(void) { 
+	mono_os_mutex_unlock (&mono_interlocked_mutex);
+}
+#endif
+
 #endif /* _MONO_METADATA_THREADS_TYPES_H_ */