!! This PR is a copy of dotnet/runtime#43893,  please do not edit or review it in this repo !!<br/>Do not automatically approve this PR:<br/><br/>* Consider how the changes affect configurations in this repo,<br/>* Check effects on files that are not mirrored,<br/>* Identify test cases that may be needed in this repo.<br/><br/>!! Merge the PR only after the original PR is merged !!<br/><hr/><br/>For microbenchmark System.Threading.Tests.Perf_Interlocked.Exchange_long, before and after this change: 141.6ns -> 11.0ns
For microbenchmark System.Threading.Tests.Perf_Volatile.Read_double, before and after this change: 183.9ns -> 158.5ns
For microbenchmark System.Threading.Tests.Perf_Volatile.Write_double, before and after this change: 151.0ns -> 18.04ns

Fixes dotnet/runtime#41449